Added navigation between screens and refactored all activities into screens
