+++ TheDialer/app/src/main/AndroidManifest.xml
@@ -12,26 +12,6 @@
         android:theme="@style/Theme.Dialer"
         tools:targetApi="31">
         <activity
-            android:name=".HistoryActivity"
-            android:exported="false"
-            android:label="@string/title_activity_history"
-            android:theme="@style/Theme.Dialer" />
-        <activity
-            android:name=".QuickDialingActivity"
-            android:exported="false"
-            android:label="@string/title_activity_quick_dialing"
-            android:theme="@style/Theme.Dialer" />
-        <activity
-            android:name=".PinView"
-            android:exported="false"
-            android:label="@string/title_activity_pin_view"
-            android:theme="@style/Theme.Dialer" />
-        <activity
-            android:name=".SendingActivity"
-            android:exported="false"
-            android:label="@string/title_activity_sending"
-            android:theme="@style/Theme.Dialer" />
-        <activity
             android:name=".MainActivity"
             android:exported="true"
             android:label="@string/app_name"
