+++ TheDialer/app/src/main/java/com/cedricbahirwe/dialer/MainActivity.kt
@@ -4,7 +4,6 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
@@ -12,18 +11,15 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import androidx.navigation.compose.rememberNavController
+import com.cedricbahirwe.dialer.navigation.NavGraph
 import com.cedricbahirwe.dialer.ui.theme.DialerTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
-            DialerTheme {
-                // A surface container using the 'background' color from the theme
-                Surface(color = MaterialTheme.colors.background) {
-                    Greeting("Android")
-                }
-            }
+            MainScreen()
         }
     }
 }
@@ -33,7 +29,13 @@ fun Greeting(name: String) {
         Text(text = "Hi, my name is $name!", modifier = Modifier.padding(24.dp))
     }
 }
-
+@Composable
+fun MainScreen() {
+    DialerTheme {
+        val navController = rememberNavController()
+        NavGraph(navController)
+    }
+}
 @Preview(showBackground = true)
 @Composable
 fun DefaultPreview() {
