+++ TheDialer/app/src/main/java/com/cedricbahirwe/dialer/screens/PinView.kt
@@ -1,8 +1,5 @@
-package com.cedricbahirwe.dialer
+package com.cedricbahirwe.dialer.screens
 
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
@@ -19,32 +16,30 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.cedricbahirwe.dialer.ui.theme.DialerTheme
 
-class PinView : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            DialerTheme {
-                // A surface container using the 'background' color from the theme
-                Surface(
-                    modifier = Modifier.fillMaxSize(),
-                    color = MaterialTheme.colors.background
-                ) {
-                    PinView("Android")
-                }
-            }
-        }
-    }
-}
 @Composable
 fun PinView(input: String, isFullMode: Boolean = false, btnSize: Float = 60f) {
     val buttons = remember {
-        listOf("1", "2", "3", "4", "5", "6", "7", "8", "9", "*", "0","#".takeIf { isFullMode } ?: "X" )
-    }
-    fun addKey(key: String) {
-        println("The new input is ${input+key}")
+        listOf(
+            "1",
+            "2",
+            "3",
+            "4",
+            "5",
+            "6",
+            "7",
+            "8",
+            "9",
+            "*",
+            "0",
+            "#".takeIf { isFullMode } ?: "X")
     }
 
-    LazyVerticalGrid(columns = GridCells.Fixed(3),
+    fun addKey(key: String) {
+        println("The new input is ${input + key}")
+    }
+
+    LazyVerticalGrid(
+        columns = GridCells.Fixed(3),
         verticalArrangement = Arrangement.spacedBy(16.dp),
         horizontalArrangement = Arrangement.spacedBy(10.dp)
     ) {
@@ -63,17 +58,20 @@ fun PinView(input: String, isFullMode: Boolean = false, btnSize: Float = 60f) {
 
 @Composable
 private fun CircleButton(title: String, size: Float, action: () -> Unit) {
-    OutlinedButton(onClick = action,
-        modifier= Modifier.size(size.dp),
+    OutlinedButton(
+        onClick = action,
+        modifier = Modifier.size(size.dp),
         shape = CircleShape,
-        border= null,
+        border = null,
         contentPadding = PaddingValues(0.dp),
         colors = ButtonDefaults.outlinedButtonColors(backgroundColor = Color.Gray.copy(0.2f))
     ) {
-        Text(text = title,
+        Text(
+            text = title,
             Modifier.padding(start = 1.dp),
             fontSize = 20.sp,
-            fontWeight = FontWeight.SemiBold)
+            fontWeight = FontWeight.SemiBold
+        )
     }
 }
 
@@ -81,6 +79,6 @@ private fun CircleButton(title: String, size: Float, action: () -> Unit) {
 @Composable
 fun PinViewPreview() {
     DialerTheme {
-        PinView("Android")
+        PinView(input = "Android")
     }
 }
